// src/app/shared/models/bic.test/calc.time.model.ts
export interface CalcTimeModel {
  background: boolean;
  segmentation: boolean;
  concept: number;
  kpi: boolean;
  association: number;
  likeDislikeMoodboard: boolean;
  likeDislikeWord: boolean;
  openFeedback: boolean;
  relevance: boolean;
  frequency: boolean;
  customQuestion: number;
  customQuestionGrid: number;
}